pyproject.toml
[tool.poetry]
name = "chat-superviser-bot"
version = "0.1.0"
description = ""
authors = ["Shishkin Alexander <shishkin.alexander@disroot.org>"]

[tool.poetry.dependencies]
python = "^3.10"
Pyrogram = "^2.0.30"
rich = "^12.4.4"
django-environ = "^0.9.0"
asyncio-redis = "^0.16.0"
TgCrypto = "^1.2.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
